# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/courses": {
      "get": {
        "operationId": "CoursesController_getCourses",
        "summary": "Get all courses",
        "description": "Retrieve courses with optional filtering and pagination",
        "parameters": [
          {
            "name": "level",
            "required": false,
            "in": "query",
            "description": "Filter by course level",
            "schema": {
              "example": "BEGINNER",
              "enum": [
                "JUNIOR",
                "SENIOR",
                "MASTER"
              ],
              "type": "string"
            }
          },
          {
            "name": "published",
            "required": false,
            "in": "query",
            "description": "Filter by published status",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "featured",
            "required": false,
            "in": "query",
            "description": "Filter by featured status",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "description": "Filter by tags (comma-separated)",
            "schema": {
              "example": "React,JavaScript",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search courses by title or description",
            "schema": {
              "example": "react hooks",
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category",
            "schema": {
              "example": "Web Development",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Courses retrieved successfully"
          }
        },
        "tags": [
          "Courses"
        ]
      },
      "post": {
        "operationId": "CoursesController_createCourse",
        "summary": "Create course",
        "description": "Create a new course. Requires admin authentication.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/courses/{id}": {
      "get": {
        "operationId": "CoursesController_getCourse",
        "summary": "Get course by ID",
        "description": "Retrieve detailed course information including chapters",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "example": "789e0123-e45b-67d8-a901-426614174222",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course retrieved successfully"
          },
          "404": {
            "description": "Course not found"
          }
        },
        "tags": [
          "Courses"
        ]
      },
      "patch": {
        "operationId": "CoursesController_updateCourse",
        "summary": "Update course",
        "description": "Update course information. Requires admin authentication.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "example": "789e0123-e45b-67d8-a901-426614174222",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Course not found"
          }
        },
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "CoursesController_deleteCourse",
        "summary": "Delete course",
        "description": "Delete a course. Requires admin authentication.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "example": "789e0123-e45b-67d8-a901-426614174222",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Course not found"
          }
        },
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/courses/{id}/chapters": {
      "get": {
        "operationId": "CoursesController_getCourseChapters",
        "summary": "Get course chapters",
        "description": "Retrieve all chapters for a course",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "example": "789e0123-e45b-67d8-a901-426614174222",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chapters retrieved successfully"
          },
          "404": {
            "description": "Course not found"
          }
        },
        "tags": [
          "Courses"
        ]
      },
      "post": {
        "operationId": "CoursesController_createChapter",
        "summary": "Create chapter",
        "description": "Create a new chapter for a course. Requires admin authentication.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "example": "789e0123-e45b-67d8-a901-426614174222",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChapterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chapter created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Course not found"
          }
        },
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "WKU Software Crew API",
    "description": "RESTful API for WKU Software Crew platform - A comprehensive learning and collaboration platform for software development students",
    "version": "1.0",
    "contact": {
      "name": "WKU Software Crew",
      "url": "https://github.com/wku-crew",
      "email": "contact@wku-crew.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Course title",
            "example": "Introduction to React",
            "minLength": 2,
            "maxLength": 200
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly course identifier",
            "example": "intro-to-react",
            "minLength": 2,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "description": "Course description",
            "example": "Learn the fundamentals of React including components, hooks, and state management",
            "minLength": 10
          },
          "thumbnail": {
            "type": "string",
            "description": "Course thumbnail image URL",
            "example": "https://images.example.com/react-course.jpg"
          },
          "level": {
            "type": "string",
            "description": "Course difficulty level",
            "enum": [
              "JUNIOR",
              "SENIOR",
              "MASTER"
            ],
            "example": "BEGINNER"
          },
          "duration": {
            "type": "number",
            "description": "Total course duration in minutes",
            "example": 180,
            "minimum": 0
          },
          "published": {
            "type": "boolean",
            "description": "Whether the course is published",
            "example": true,
            "default": false
          },
          "featured": {
            "type": "boolean",
            "description": "Whether the course is featured",
            "example": false,
            "default": false
          },
          "tags": {
            "description": "Course tags for categorization",
            "example": [
              "React",
              "JavaScript",
              "Frontend"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "category": {
            "type": "string",
            "description": "Course category",
            "example": "Web Development"
          }
        },
        "required": [
          "title",
          "slug",
          "description",
          "level",
          "duration"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Course title",
            "example": "Advanced React Patterns",
            "minLength": 2,
            "maxLength": 200
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly course identifier",
            "example": "advanced-react-patterns",
            "minLength": 2,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "description": "Course description",
            "example": "Deep dive into advanced React patterns and best practices",
            "minLength": 10
          },
          "thumbnail": {
            "type": "string",
            "description": "Course thumbnail image URL",
            "example": "https://images.example.com/advanced-react.jpg"
          },
          "level": {
            "type": "string",
            "description": "Course difficulty level",
            "enum": [
              "JUNIOR",
              "SENIOR",
              "MASTER"
            ],
            "example": "ADVANCED"
          },
          "duration": {
            "type": "number",
            "description": "Total course duration in minutes",
            "example": 240,
            "minimum": 0
          },
          "published": {
            "type": "boolean",
            "description": "Whether the course is published",
            "example": true
          },
          "featured": {
            "type": "boolean",
            "description": "Whether the course is featured",
            "example": true
          },
          "tags": {
            "description": "Course tags",
            "example": [
              "React",
              "Advanced",
              "JavaScript"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "category": {
            "type": "string",
            "description": "Course category",
            "example": "Advanced Web Development"
          }
        }
      },
      "CreateChapterDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Chapter title",
            "example": "1. Introduction to React"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug",
            "example": "introduction-to-react"
          },
          "order": {
            "type": "number",
            "description": "Order in the course",
            "example": 1
          },
          "duration": {
            "type": "number",
            "description": "Duration in minutes",
            "example": 15
          },
          "videoUrl": {
            "type": "string",
            "description": "Video URL (YouTube, Vimeo, etc.)"
          },
          "content": {
            "type": "string",
            "description": "Chapter content in Markdown"
          }
        },
        "required": [
          "title",
          "slug",
          "order"
        ]
      },
...
    }
  }
}